Feat/ adding orderbook summary hash support
This PR aims to add the orderbook summary struct type and to provide a helper to calculate the hash of an orderbook.

To generate the hash of an orderbook it is needed a `OrderBookSummary` object.

This can be fetched from CLOB:

```ts
const orderbook = await clobClient.getOrderBook("token id" )
```

Or built manually:

```ts
const orderbook = {
     market: "0x...",
     asset_id: "123...",
     bids: [],
     asks: [],
} as OrderBookSummary;
```

Then to generate the hash (if we pull the orderbook from clob the hash is included in the response) there are two options

```python
# using the client
let hash = clobClient.getOrderBookHash(orderbook);

# or using the util func
hash = generateOrderBookSummaryHash(orderbook)
```